requirements.txt
-i https://pypi.org/simple
certifi==2023.5.7; python_version >= '3.6'
charset-normalizer==3.1.0; python_full_version >= '3.7.0'
future==0.18.3; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4; python_version >= '3.5'
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
oauthlib==3.2.2; python_version >= '3.6'
pillow==9.5.0; python_version >= '3.7'
pygments==2.17.2; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
pytumblr==0.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==13.7.0; python_full_version >= '3.7.0'
urllib3==2.0.3; python_version >= '3.7'
